revert(player): drop everything outside the review list
Reverts the desktop affordances and the display settings from this branch,
per review. They will be proposed separately, one per PR, so each can be
judged on its own rather than bundled into a fix-up.

Removed: the scrub ring around the artwork and its geometry module, the
hover rail, the docked chapters and bookmarks panels, the sleep timer
popover, the keyboard shortcuts sheet and the hotkeys it documented, the
playback rate pill, the white play button, the pill-shaped track bars,
the three-column bottom bar, and every player setting added here.

The shared ChapterList / BookmarkList / SleepTimerPanel extractions go
with them: their second consumer was the docked panels, so on their own
they would be churn in a diff that should stay small.

Also restores the chapters icon to format_list_bulleted, matching the
mini player - that was a regression, caught in review.

What remains is the review list and nothing else: the back button closing
full-screen, Escape scoped to popovers, an inert background, the landscape
layout, the volume readout no longer duplicating the popover, distinct
slider labels, the artwork not painting before it is measured, the queue
button box, and the enter/exit transitions.
