+++ src/assets/transitions.css
@@ -133,47 +133,6 @@
     opacity 250ms ease-in;
 }
 
-/* Fullscreen hover rail — controls beside the artwork, staggered out from behind it.
-   Only opacity and transform move: pointer-events are never touched, so a control that is
-   painted is always clickable. The rail is rendered by the desktop layout alone, which is
-   already gated on the primary input being able to hover, so the media query here is a
-   second belt rather than the only one. */
-@media (hover: hover) and (pointer: fine) {
-  .player-fab-stack > * {
-    opacity: 0;
-    transform: translateX(-14px) scale(0.9);
-    transition:
-      opacity 0.18s ease,
-      transform 0.28s cubic-bezier(0.34, 1.4, 0.64, 1);
-  }
-
-  .group:hover .player-fab-stack > *,
-  .group:focus-within .player-fab-stack > * {
-    opacity: 1;
-    transform: translateX(0) scale(1);
-  }
-
-  .group:hover .player-fab-stack > *:nth-child(1),
-  .group:focus-within .player-fab-stack > *:nth-child(1) {
-    transition-delay: 0.02s;
-  }
-
-  .group:hover .player-fab-stack > *:nth-child(2),
-  .group:focus-within .player-fab-stack > *:nth-child(2) {
-    transition-delay: 0.06s;
-  }
-
-  .group:hover .player-fab-stack > *:nth-child(3),
-  .group:focus-within .player-fab-stack > *:nth-child(3) {
-    transition-delay: 0.1s;
-  }
-
-  .group:hover .player-fab-stack > *:nth-child(4),
-  .group:focus-within .player-fab-stack > *:nth-child(4) {
-    transition-delay: 0.14s;
-  }
-}
-
 /* Fullscreen player: rises out of the mini player and drops back into it.
    `player-fullscreen-exit` must stay in step with FULLSCREEN_EXIT_MS. */
 .player-fullscreen-enter {
@@ -230,49 +189,6 @@
 }
 
 @keyframes player-fullscreen-artwork-in {
-  from {
-    opacity: 0;
-    transform: scale(0.94);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-/* Playback rate pill: springs in when the rate leaves 1x, and nudges on each step.
-   Transition-driven rather than a keyframe, so a held key holds the scale instead of
-   restarting an animation mid-flight. */
-.player-speed-pill {
-  animation: player-speed-pill-in 0.32s cubic-bezier(0.34, 1.45, 0.64, 1);
-  transition:
-    background-color 0.15s ease,
-    transform 0.18s cubic-bezier(0.34, 1.45, 0.64, 1);
-}
-
-.player-speed-pill-pulse {
-  transform: scale(1.12);
-}
-
-@keyframes player-speed-pill-in {
-  from {
-    opacity: 0;
-    transform: translateY(6px) scale(0.88);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-  }
-}
-
-/* Volume popover: grows into place rather than appearing whole. Scale only, no directional
-   offset — the same panel opens above the trigger in the toolbars and beside it on the
-   fullscreen rail, and a translate that suits one reads backwards in the other. */
-.player-volume-popover {
-  animation: player-volume-popover-in 0.16s ease-out;
-}
-
-@keyframes player-volume-popover-in {
   from {
     opacity: 0;
     transform: scale(0.94);
@@ -386,10 +302,6 @@
 }
 
 @media (prefers-reduced-motion: reduce) {
-  .player-fab-stack > * {
-    transition: none;
-  }
-
   .player-fullscreen-enter,
   .player-fullscreen-pane,
   .player-fullscreen-artwork {
@@ -406,14 +318,8 @@
     }
   }
 
-  .player-volume-hud,
-  .player-volume-popover,
-  .player-speed-pill {
+  .player-volume-hud {
     animation: none;
-  }
-
-  .player-speed-pill-pulse {
-    transform: none;
   }
 
   .player-volume-hud-fill {
