+++ src/components/player/PlayerTrackBar.tsx
@@ -21,18 +21,6 @@ interface PlayerTrackBarProps {
   hideChapterTitle?: boolean
   /** Overrides the slider's accessible name — stacked bars must not share one */
   ariaLabel?: string
-  /**
-   * Taller pill-shaped bar with rounded ends, for the fullscreen player. Ports the Vue
-   * client's `rounded` variant, where the same bar carries the whole layout rather than
-   * sitting in a dense strip.
-   */
-  rounded?: boolean
-  /**
-   * `stacked` puts the timestamps and chapter under the bar, which is what the mini player
-   * has room for. `flanking` puts the chapter above and a timestamp at each end of the bar,
-   * the way the Vue client's fullscreen player lays it out.
-   */
-  layout?: 'stacked' | 'flanking'
   className?: string
 }
 
@@ -47,8 +35,6 @@ export default function PlayerTrackBar({
   scope = 'auto',
   hideChapterTitle = false,
   ariaLabel,
-  rounded = false,
-  layout = 'stacked',
   className
 }: PlayerTrackBarProps) {
   const t = useTypeSafeTranslations()
@@ -270,169 +256,123 @@ export default function PlayerTrackBar({
 
   const isMobileCollapsed = variant === 'mobile-collapsed'
 
-  const barBlock = (
-    <div className="relative">
-      {/* The padding widens the touch target without changing layout height, which the
-            negative margin cancels. The visual bar stays 8px. */}
-      <div
-        role="slider"
-        tabIndex={0}
-        aria-label={ariaLabel ?? t('LabelPlaybackPosition')}
-        aria-valuemin={0}
-        aria-valuemax={100}
-        aria-valuenow={Math.round(playedPercent)}
-        aria-valuetext={`${currentTimeFormatted} / ${timeRemainingFormatted}`}
-        className="group/track relative -my-2.5 w-full cursor-pointer touch-none py-2.5"
-        onPointerDown={handlePointerDown}
-        onPointerMove={handlePointerMove}
-        onPointerUp={handlePointerUp}
-        onPointerCancel={handlePointerUp}
-        onPointerLeave={handlePointerLeave}
-        onKeyDown={handleKeyDown}
-      >
-        <div
-          ref={trackRef}
-          className={mergeClasses(
-            'bg-track-bg relative w-full transition-transform duration-100 group-hover/track:scale-y-125',
-            // The rounded variant still clips, or the square-ended fills would poke out of
-            // the pill at both ends
-            rounded ? 'h-2.5 overflow-hidden rounded-full' : 'h-2 overflow-hidden'
-          )}
-        >
-          {/* HLS transcode ready track (server-side segment progress) */}
-          {isHlsTranscode && (
-            <div
-              className="bg-track-progress/30 pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75"
-              style={{ width: `${transcodeReadyPercent}%` }}
-            />
-          )}
-          {/* Buffer track */}
-          <div
-            className="bg-track-progress/50 pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75"
-            style={{ width: `${bufferedPercent}%` }}
-          />
-          {/* Played track */}
-          <div
-            className={mergeClasses(
-              'bg-track-progress pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75',
-              rounded ? 'rounded-full' : ''
-            )}
-            style={{ width: `${playedPercent}%` }}
-          />
-          {/* Track cursor (vertical line on hover) */}
-          <div
-            ref={trackCursorRef}
-            className={mergeClasses(
-              'bg-track-progress pointer-events-none absolute top-0 left-0 h-full w-0.5 transition-opacity duration-100',
-              isHovering ? 'opacity-100' : 'opacity-0'
-            )}
-          />
-          {/* Loading animation - sliding shimmer effect */}
-          {isLoading && (
-            <div className="via-track-progress/30 loading-track-slide pointer-events-none absolute top-0 h-full w-1/4 bg-gradient-to-r from-transparent to-transparent" />
-          )}
-        </div>
-      </div>
-
-      {/* Chapter ticks */}
-      <div className={mergeClasses('relative h-2 w-full overflow-hidden', useChapterTrack ? 'opacity-0' : '')}>
-        {chapterTicks.map((tick, index) => (
-          <div key={index} className="bg-track-progress/30 pointer-events-none absolute top-0 h-1 w-px" style={{ left: `${tick.left}px` }} />
-        ))}
-      </div>
-
-      {/* Hover timestamp */}
-      <div
-        ref={hoverTimestampRef}
-        className={mergeClasses(
-          'bg-foreground text-background pointer-events-none absolute -top-8 left-0 z-10 rounded-full transition-opacity duration-100',
-          isHovering ? 'opacity-100' : 'opacity-0'
-        )}
-      >
-        <p ref={hoverTimestampTextRef} className="truncate px-2 py-0.5 text-center font-mono text-xs whitespace-nowrap">
-          00:00
-        </p>
-      </div>
-
-      {/* Hover timestamp arrow */}
-      <div
-        ref={hoverTimestampArrowRef}
-        className={mergeClasses(
-          'bg-foreground text-background pointer-events-none absolute -top-3.5 left-0 rounded-full transition-opacity duration-100',
-          isHovering ? 'opacity-100' : 'opacity-0'
-        )}
-      >
-        <div className="absolute right-0 -bottom-1.5 left-0 flex w-full justify-center">
-          <div className="border-t-foreground h-0 w-0 border-t-4 border-r-4 border-l-4 border-r-transparent border-l-transparent" />
-        </div>
-      </div>
-    </div>
-  )
-
-  // Fullscreen: chapter above the bar, the two timestamps flanking it. The timestamps read
-  // as the ends of the bar this way instead of as a caption under it, and the chapter gets
-  // the full width rather than whatever is left between two strings of unequal length.
-  if (layout === 'flanking') {
-    return (
-      <div className={className}>
-        {!hideChapterTitle && (
-          <div className="mb-2 flex items-center justify-center gap-2">
-            {/* A plain truncating span, not TruncatingTooltipText: that wrapper is `w-full`, which
-                as a flex item eats the whole row and pushes the percentage out to the edge
-                instead of letting the pair sit centred together */}
-            {currentChapter ? (
-              <span dir="auto" className="text-foreground-muted min-w-0 truncate text-sm">
-                {currentChapter.title}
-              </span>
-            ) : (
-              <span className="text-foreground-subdued text-sm">{t('LabelChapter')}</span>
-            )}
-            <span className="text-foreground-subdued shrink-0 font-mono text-xs">· {Math.round(playedPercent)}%</span>
-          </div>
-        )}
-
-        {/* -mt-1 offsets the chapter tick strip the bar renders below itself, so the
-            timestamps sit level with the bar rather than with the whole block */}
-        <div className="flex items-center gap-2.5">
-          <p className="text-foreground-muted -mt-1 w-24 shrink-0 text-end font-mono text-base leading-none">{currentTimeFormatted}</p>
-          <div className="min-w-0 grow">{barBlock}</div>
-          <p className="text-foreground-muted -mt-1 w-24 shrink-0 font-mono text-base leading-none">{timeRemainingFormatted}</p>
-        </div>
-      </div>
-    )
-  }
-
   return (
     <div className={className}>
-      {barBlock}
-      {/* Both timestamps get the same fixed width, so the chapter title between them is centred
-          on the bar rather than on whatever is left after two strings of different lengths */}
-      <div className={mergeClasses('flex items-center gap-3', isMobileCollapsed ? 'mt-0.5' : '')}>
-        <p className={mergeClasses('text-foreground-muted w-24 shrink-0 text-start font-mono', isMobileCollapsed ? 'text-xs' : 'text-sm')}>
+      <div className="relative">
+        {/* The padding widens the touch target without changing layout height, which the
+            negative margin cancels. The visual bar stays 8px. */}
+        <div
+          role="slider"
+          tabIndex={0}
+          aria-label={ariaLabel ?? t('LabelPlaybackPosition')}
+          aria-valuemin={0}
+          aria-valuemax={100}
+          aria-valuenow={Math.round(playedPercent)}
+          aria-valuetext={`${currentTimeFormatted} / ${timeRemainingFormatted}`}
+          className="group/track relative -my-2.5 w-full cursor-pointer touch-none py-2.5"
+          onPointerDown={handlePointerDown}
+          onPointerMove={handlePointerMove}
+          onPointerUp={handlePointerUp}
+          onPointerCancel={handlePointerUp}
+          onPointerLeave={handlePointerLeave}
+          onKeyDown={handleKeyDown}
+        >
+          <div ref={trackRef} className="bg-track-bg relative h-2 w-full overflow-hidden transition-transform duration-100 group-hover/track:scale-y-125">
+            {/* HLS transcode ready track (server-side segment progress) */}
+            {isHlsTranscode && (
+              <div
+                className="bg-track-progress/30 pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75"
+                style={{ width: `${transcodeReadyPercent}%` }}
+              />
+            )}
+            {/* Buffer track */}
+            <div
+              className="bg-track-progress/50 pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75"
+              style={{ width: `${bufferedPercent}%` }}
+            />
+            {/* Played track */}
+            <div
+              className="bg-track-progress pointer-events-none absolute top-0 left-0 h-full transition-[width] duration-75"
+              style={{ width: `${playedPercent}%` }}
+            />
+            {/* Track cursor (vertical line on hover) */}
+            <div
+              ref={trackCursorRef}
+              className={mergeClasses(
+                'bg-track-progress pointer-events-none absolute top-0 left-0 h-full w-0.5 transition-opacity duration-100',
+                isHovering ? 'opacity-100' : 'opacity-0'
+              )}
+            />
+            {/* Loading animation - sliding shimmer effect */}
+            {isLoading && (
+              <div className="via-track-progress/30 loading-track-slide pointer-events-none absolute top-0 h-full w-1/4 bg-gradient-to-r from-transparent to-transparent" />
+            )}
+          </div>
+        </div>
+
+        {/* Chapter ticks */}
+        <div className={mergeClasses('relative h-2 w-full overflow-hidden', useChapterTrack ? 'opacity-0' : '')}>
+          {chapterTicks.map((tick, index) => (
+            <div key={index} className="bg-track-progress/30 pointer-events-none absolute top-0 h-1 w-px" style={{ left: `${tick.left}px` }} />
+          ))}
+        </div>
+
+        {/* Hover timestamp */}
+        <div
+          ref={hoverTimestampRef}
+          className={mergeClasses(
+            'bg-foreground text-background pointer-events-none absolute -top-8 left-0 z-10 rounded-full transition-opacity duration-100',
+            isHovering ? 'opacity-100' : 'opacity-0'
+          )}
+        >
+          <p ref={hoverTimestampTextRef} className="truncate px-2 py-0.5 text-center font-mono text-xs whitespace-nowrap">
+            00:00
+          </p>
+        </div>
+
+        {/* Hover timestamp arrow */}
+        <div
+          ref={hoverTimestampArrowRef}
+          className={mergeClasses(
+            'bg-foreground text-background pointer-events-none absolute -top-3.5 left-0 rounded-full transition-opacity duration-100',
+            isHovering ? 'opacity-100' : 'opacity-0'
+          )}
+        >
+          <div className="absolute right-0 -bottom-1.5 left-0 flex w-full justify-center">
+            <div className="border-t-foreground h-0 w-0 border-t-4 border-r-4 border-l-4 border-r-transparent border-l-transparent" />
+          </div>
+        </div>
+      </div>
+      <div className={mergeClasses('flex items-center justify-between gap-3', isMobileCollapsed ? 'mt-0.5' : '')}>
+        <p className={mergeClasses('text-foreground-muted shrink-0 font-mono', isMobileCollapsed ? 'text-xs' : 'text-sm')}>
           {currentTimeFormatted}
           {' / '}
           {Math.round(playedPercent)}%
         </p>
         {currentChapter && !hideChapterTitle ? (
-          <div className="text-foreground-muted flex min-w-0 flex-1 items-center justify-center">
-            <TruncatingTooltipText
-              lazy
-              text={currentChapter.title}
-              className={mergeClasses('min-w-0', isMobileCollapsed ? 'text-xs' : 'text-sm')}
-              position="top"
-            />
-            {useChapterTrack && currentChapterNumber !== null && (
-              <span className="text-foreground-subdued shrink-0 pl-1 text-xs">
-                ({currentChapterNumber} of {chapters.length})
-              </span>
-            )}
-          </div>
+          isMobileCollapsed ? (
+            <div className="text-foreground-muted flex min-w-0 flex-1 items-center justify-center sm:max-w-none">
+              <TruncatingTooltipText lazy text={currentChapter.title} className="min-w-0 text-xs" position="top" />
+              {useChapterTrack && currentChapterNumber !== null && (
+                <span className="text-foreground-subdued shrink-0 pl-1 text-xs">
+                  ({currentChapterNumber} of {chapters.length})
+                </span>
+              )}
+            </div>
+          ) : (
+            <p className="text-foreground-muted max-w-[40%] truncate text-sm sm:max-w-none">
+              {currentChapter.title}{' '}
+              {useChapterTrack && (
+                <span className="text-foreground-subdued pl-1 text-xs">
+                  ({currentChapterNumber} of {chapters.length})
+                </span>
+              )}
+            </p>
+          )
         ) : (
           <span className="flex-1" />
         )}
-        <p className={mergeClasses('text-foreground-muted w-24 shrink-0 text-end font-mono', isMobileCollapsed ? 'text-xs' : 'text-sm')}>
-          {timeRemainingFormatted}
-        </p>
+        <p className={mergeClasses('text-foreground-muted shrink-0 font-mono', isMobileCollapsed ? 'text-xs' : 'text-sm')}>{timeRemainingFormatted}</p>
       </div>
     </div>
   )
