+++ src/locales/en-us.json
@@ -196,7 +196,6 @@
   "HeaderIgnoredFiles": "Ignored Files",
   "HeaderItemFiles": "Item Files",
   "HeaderItemMetadataUtils": "Item Metadata Utils",
-  "HeaderKeyboardShortcuts": "Keyboard Shortcuts",
   "HeaderLastListeningSession": "Last Listening Session",
   "HeaderLatestEpisodes": "Latest episodes",
   "HeaderLibraries": "Libraries",
@@ -225,7 +224,6 @@
   "HeaderOtherFiles": "Other Files",
   "HeaderPasswordAuthentication": "Password Authentication",
   "HeaderPermissions": "Permissions",
-  "HeaderPlayback": "Playback",
   "HeaderPlayerQueue": "Player Queue",
   "HeaderPlayerSettings": "Player Settings",
   "HeaderPlaylist": "Playlist",
@@ -300,7 +298,6 @@
   "LabelAllUsersExcludingGuests": "All users excluding guests",
   "LabelAllUsersIncludingGuests": "All users including guests",
   "LabelAlreadyInYourLibrary": "Already in your library",
-  "LabelAmoledPlayerSurfaces": "AMOLED black player surfaces",
   "LabelApiKeyCreated": "API Key \"{0}\" created successfully.",
   "LabelApiKeyCreatedDescription": "Make sure to copy the API key now as you will not be able to see this again.",
   "LabelApiKeyUser": "Act on behalf of user",
@@ -322,7 +319,6 @@
   "LabelAutoFetchMetadataHelp": "Fetches metadata for title, author, and series to streamline uploading. Additional metadata may have to be matched after upload.",
   "LabelAutoLaunch": "Auto Launch",
   "LabelAutoLaunchDescription": "Redirect to the auth provider automatically when navigating to the login page (manual override path <code>/login?autoLaunch=0</code>)",
-  "LabelAutoOpenFullscreenOnPlay": "Open fullscreen player when playback starts",
   "LabelAutoPlay": "Auto Play",
   "LabelAutoRegister": "Auto Register",
   "LabelAutoRegisterDescription": "Automatically create new users after logging in",
@@ -343,7 +339,6 @@
   "LabelByAuthor": "by {0}",
   "LabelChangePassword": "Change Password",
   "LabelChannels": "Channels",
-  "LabelChapter": "Chapter",
   "LabelChapterCount": "{0} Chapters",
   "LabelChapterTitle": "Chapter Title",
   "LabelChapters": "Chapters",
@@ -672,9 +667,7 @@
   "LabelPlaybackPositionInBook": "Playback position in book",
   "LabelPlaybackPositionInChapter": "Playback position in chapter",
   "LabelPlaybackRate": "Playback Rate",
-  "LabelPlaybackRateFaster": "Faster",
   "LabelPlaybackRateIncrementDecrement": "Playback Rate Increment/Decrement Amount",
-  "LabelPlaybackRateSlower": "Slower",
   "LabelPlayerChapterNumberMarker": "{0} of {1}",
   "LabelPlayerFullscreen": "Fullscreen player",
   "LabelPlaylists": "Playlists",
@@ -793,14 +786,8 @@
   "LabelShareOpen": "Share Open",
   "LabelShareURL": "Share URL",
   "LabelShowAll": "Show All",
-  "LabelShowBookTrackWithChapterTrack": "Show book track alongside chapter track",
-  "LabelShowBookmarksInPlayerBar": "Show bookmarks in the player bar",
-  "LabelShowCoverProgressRing": "Show progress ring around cover",
-  "LabelShowFullscreenCornerButtons": "Show fullscreen corner buttons",
-  "LabelShowQueueButton": "Show queue button",
   "LabelShowSeconds": "Show seconds",
   "LabelShowSubtitles": "Show Subtitles",
-  "LabelShowVolumeInPlayerBar": "Show volume in the player bar",
   "LabelSigningAlgorithm": "Signing Algorithm",
   "LabelSize": "Size",
   "LabelSizeWithValue": "Size: {0}",
@@ -905,22 +892,17 @@
   "LabelUseAdvancedOptions": "Use Advanced Options",
   "LabelUseChapterTrack": "Use chapter track",
   "LabelUseFullTrack": "Use full track",
-  "LabelUseLegacyBookmarksDialog": "Use legacy bookmarks dialog",
-  "LabelUseLegacySleepTimerDialog": "Use legacy sleep timer dialog",
   "LabelUseZeroForUnlimited": "Use 0 for unlimited",
   "LabelUser": "User",
   "LabelUserinfoURL": "Userinfo URL",
   "LabelUsername": "Username",
   "LabelValue": "Value",
   "LabelVersion": "Version",
-  "LabelView": "View",
   "LabelViewBookmarks": "View bookmarks",
   "LabelViewChapters": "View chapters",
   "LabelViewPlayerSettings": "View player settings",
   "LabelViewQueue": "View player queue",
   "LabelVolume": "Volume",
-  "LabelVolumeDown": "Volume down",
-  "LabelVolumeUp": "Volume up",
   "LabelWarning": "Warning",
   "LabelWebRedirectURLsDescription": "Authorize these URLs in your OAuth provider to allow redirection back to the web app after login:",
   "LabelWebRedirectURLsSubfolder": "Subfolder for Redirect URLs",
